feat: Officially support matrix builds
BREAKING CHANGE: The `publish` command was removed and merged with
the `release` command. To better support projects with multiple
builds (e.g. against multiple versions of node) the release parts
were split from the verification parts. `nlm verify` can be used
as a `posttest` script in those cases. This ensures that multiple
builds won't try to release the same version.