feat: re-check publish status for releaseType none
If `npm publish` fails, you're left with a version that nlm believes is
real (has tag, CHANGELOG, GH Release, etc.), but that doesn't exist on
the npm registry.  Restarting the CI job doesn't help, because you
either get an attempted dup commit, or the changes are "none" so it
aborts.

This change makes it so that when you re-run a CI job for a version
bump/CHANGELOG commit, it will double-check that the current version
is published, and try (again) to publish it if it's not.